Move ArgList out of types.h
